// web/src/shared/api/generated/model/planPriceResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { BillingPeriod } from './billingPeriod';
import type { Money } from './money';

export interface PlanPriceResult {
  period: BillingPeriod;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  months: number | string;
  price: Money;
}